timestamps have "not null" contraints by default in rails3.
The records table needs to have the "not null" constraints removed for the three timestamps (created_at, updated_at, change_date) to support SLAVE domains.
pdns has no knowledge of the fields, leaves them undefined when inserting records.

Additionally the schema.rb was updated to to reflect the current state of the migrations.